move toggle theme logic out of App.vue
