+++ src/components/ThemeToggle.vue
@@ -1,19 +1,25 @@
 <template>
-    <label class="flex cursor-pointer gap-2">
-      <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="5" /><path d="M12 1v2M12 21v2M4.2 4.2l1.4 1.4M18.4 18.4l1.4 1.4M1 12h2M21 12h2M4.2 19.8l1.4-1.4M18.4 5.6l1.4-1.4" /></svg>
-      <input type="checkbox" :checked="theme === 'dracula'" @change="$emit('toggle-theme')" class="toggle theme-controller" />
-      <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z"></path></svg>
-    </label>
-  </template>
-  
-  <script setup>
-  defineProps({
-    theme: {
-      type: String,
-      required: true
-    }
-  })
-  
-  defineEmits(['toggle-theme'])
-  </script>
-  
+  <label class="flex cursor-pointer gap-2">
+    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="5" /><path d="M12 1v2M12 21v2M4.2 4.2l1.4 1.4M18.4 18.4l1.4 1.4M1 12h2M21 12h2M4.2 19.8l1.4-1.4M18.4 5.6l1.4-1.4" /></svg>
+    <input type="checkbox" :checked="theme === 'dracula'" @change="toggleTheme" class="toggle theme-controller" />
+    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z"></path></svg>
+  </label>
+</template>
+
+<script setup>
+import { ref, onMounted } from 'vue';
+
+const theme = ref(localStorage.getItem('theme') || 'emerald')
+const emit = defineEmits(['theme-changed'])
+
+const toggleTheme = () => {
+  theme.value = theme.value === 'emerald' ? 'dracula' : 'emerald'
+  localStorage.setItem('theme', theme.value)
+  document.documentElement.setAttribute('data-theme', theme.value)
+  emit('theme-changed', theme.value)
+}
+
+onMounted(() => {
+  document.documentElement.setAttribute('data-theme', theme.value)
+})
+</script>
